[redhat] Add a script to automatically rebase ark-patches
This script rebases the ark-patches against the latest master branch. If
patches cannot be rebased, issues are filed in the GitLab issue tracker.

This script is really only intended to be run in CI on a regular
schedule (nightly, most likely). A sample .gitlab-ci.yml would be:

ark_patches_rebase:
  script:
    - dnf -y install python3-gitlab
    - git config user.name "CKI@GitLab"
    - git config user.email "cki-project@redhat.com"
    - redhat/scripts/ci/ark-rebase-patches.sh "master" $CI_PROJECT_ID
    - git push -f origin ark-patches

Signed-off-by: Jeremy Cline <jcline@redhat.com>