New configs in drivers/misc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_UACCE:

 UACCE provides interface for the user process to access the hardware
 without interaction with the kernel space in data path.

 The user-space interface is described in
 include/uapi/misc/uacce/uacce.h

 See Documentation/misc-devices/uacce.rst for more details.

 If you don't know what to do here, say N.

 Symbol: UACCE [=n]
 Type  : tristate
 Defined at drivers/misc/uacce/Kconfig:1
   Prompt: Accelerator Framework for User Land
   Depends on: IOMMU_API [=y]
   Location:
     -> Device Drivers
       -> Misc devices

Cc: John Linville <linville@redhat.com>