[redhat] New configs in drivers/net/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_BAREUDP:

 This adds a bare UDP tunnel module for tunnelling different
 kinds of traffic like MPLS, IP, etc. inside a UDP tunnel.

 To compile this driver as a module, choose M here: the module
 will be called bareudp.

 Symbol: BAREUDP [=n]
 Type  : tristate
 Defined at drivers/net/Kconfig:262
   Prompt: Bare UDP Encapsulation
   Depends on: NETDEVICES [=y] && NET_CORE [=y] && INET [=y] && (IPV6 [=y] || !IPV6 [=y])
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> Network core driver support (NET_CORE [=y])
 Selects: NET_UDP_TUNNEL [=m] && GRO_CELLS [=y]

Enabled at the request of Jiri Benc

---

Cc: Jiri Benc <jbenc@redhat.com>