[redhat] New configs in kernel/power
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_HIBERNATION_SNAPSHOT_DEV:

 Device used by the uswsusp tools.

 Say N if no snapshotting from userspace is needed, this also
 reduces the attack surface of the kernel.

 If in doubt, say Y.

 Symbol: HIBERNATION_SNAPSHOT_DEV [=y]
 Type  : bool
 Defined at kernel/power/Kconfig:83
   Prompt: Userspace snapshot device
   Depends on: HIBERNATION [=y]
   Location:
     -> Power management and ACPI options
       -> Hibernation (aka 'suspend to disk') (HIBERNATION [=y])

---

Signed-off-by: CKI@GitLab <cki-project@redhat.com>