New configs in drivers/net/phy
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_BCM54140_PHY:

 Support the Broadcom BCM54140 Quad SGMII/QSGMII PHY.

 This driver also supports the hardware monitoring of this PHY and
 exposes voltage and temperature sensors.

 Symbol: BCM54140_PHY [=n]
 Type  : tristate
 Defined at drivers/net/phy/Kconfig:357
   Prompt: Broadcom BCM54140 PHY
   Depends on: NETDEVICES [=y] && PHYLIB [=y] && (HWMON [=y] || HWMON [=y]=n)
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> PHY Device support and infrastructure (PHYLIB [=y])
 Selects: BCM_NET_PHYLIB [=m]

Cc: Petr Oros <poros@redhat.com>