[redhat] New configs in sound/soc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_SND_SOC_AMD_RENOIR_MACH:

 This option enables machine driver for DMIC

 Symbol: SND_SOC_AMD_RENOIR_MACH [=n]
 Type  : tristate
 Defined at sound/soc/amd/Kconfig:47
   Prompt: AMD Renoir support for DMIC
   Depends on: SOUND [=m] && !UML && SND [=m] && SND_SOC [=m] && SND_SOC_AMD_RENOIR [=m]
   Location:
     -> Device Drivers
       -> Sound card support (SOUND [=m])
         -> Advanced Linux Sound Architecture (SND [=m])
           -> ALSA for SoC audio support (SND_SOC [=m])
             -> AMD Audio Coprocessor - Renoir support (SND_SOC_AMD_RENOIR [=m])
 Selects: SND_SOC_DMIC [=m]

---

Cc: Jaroslav Kysela <jkysela@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>