configs: enable BPF LSM on Fedora and ARK
We are going to want to enable BPF LSM in RHEL, so start building it on
ARK and update the default LSM list so that it is enabled on boot by
default (to enable testing). It should have almost no performance impact
unless userspace attaches some BPF programs to the hooks. Users can
still completely turn it off by adding lsm="yama,integrity,selinux" to
the kernel command line.

Signed-off-by: Ondrej Mosnacek <omosnace@redhat.com>