Create Makefile.variables for a single point of configuration change
The existing Makefiles and scripts are hard coded to expect that we are
building from the os-build branch with master being the upstream branch.
This does not work well for fedora-stable releases, and does not offer
flexibility to build different trees without knowing the Makefiles and
scripts well enough to know what to change.  This change creates a
redhat/Makefile.variables which should be a self documented single file
to change when building with a different upstream or dist branch.

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>