Turn CONFIG_DEVMEM back off for aarch64
We have ended up with a config mismatch for the RHEL configs in a way
that was unclear and untracked. CONFIG_STRICT_DEVMEM is enabled in
configs/common/generic, but CONFIG_DEVMEM was disabled for aarch64 in
RHEL (though not in Fedora). This creates a mismatch situation where the
configs do not get generated in the way they are set.  Our options are
to either turn off CONFIG_STRICT_DEVMEM as I have here, or turn on
CONFIG_DEVMEM for aarch64.  I went with the former after discussion in
the MR.

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>