[redhat] New configs in kernel/rcu
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: RCU_NOCB_CPU_DEFAULT_ALL [=n]
 Type  : bool
 Defined at kernel/rcu/Kconfig:267
   Prompt: Offload RCU callback processing from all CPUs by default
   Depends on: RCU_NOCB_CPU [=y]
   Location:
     Main menu
       -> General setup
         -> RCU Subsystem
           -> Offload RCU callback processing from boot-selected CPUs (RCU_NOCB_CPU [=y])

---

Cc: Waiman Long <longman@redhat.com>
Cc: Artem Savkov <asavkov@redhat.com>
Cc: Clark Williams <williams@redhat.com>
Cc: Donghai Qiao <dqiao@redhat.com>
Cc: Phil Auld <pauld@redhat.com>
Cc: Wander Costa <wander@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>