[redhat] New configs in net/rxrpc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: RXPERF [=n]
 Type  : tristate
 Defined at net/rxrpc/Kconfig:61
   Prompt: RxRPC test service
   Depends on: NET [=y] && AF_RXRPC [=m]
   Location:
     -> Networking support (NET [=y])
       -> RxRPC session sockets (AF_RXRPC [=m])
         -> RxRPC test service (RXPERF [=n])

---

Cc: Marc Dionne <mdionne@redhat.com>
Cc: Alice Mitchell <ajmitchell@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_RXPERF enabled as a module for ark.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>