[redhat] New configs in drivers/input
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: TOUCHSCREEN_NOVATEK_NVT_TS [=n]
 Type  : tristate
 Defined at drivers/input/touchscreen/Kconfig:657
   Prompt: Novatek NVT-ts touchscreen support
   Depends on: INPUT [=y] && INPUT_TOUCHSCREEN [=y] && I2C [=y]
   Location:
     -> Device Drivers
       -> Input device support
         -> Generic input layer (needed for keyboard, mouse, ...) (INPUT [=y])
           -> Touchscreens (INPUT_TOUCHSCREEN [=y])
             -> Novatek NVT-ts touchscreen support (TOUCHSCREEN_NOVATEK_NVT_TS [=n])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>