[redhat] New configs in arch/x86
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: INTEL_TDX_HOST [=n]
 Type  : bool
 Defined at arch/x86/Kconfig:1943
   Prompt: Intel Trust Domain Extensions (TDX) host support
   Depends on: CPU_SUP_INTEL [=y] && X86_64 [=y] && KVM_INTEL [=m]
   Location:
     -> Processor type and features
       -> Intel Trust Domain Extensions (TDX) host support (INTEL_TDX_HOST [=n])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>