bump the sensitivity of AMD SFH
Bumps the sensitivity of AMD sfh gyro and accelerometers by removing the division
operation and rebasing the units in the hid descriptor.
This helps with the gyro deadzone of the Legion Go. Should not affect existing
devices.

Since most of the magic happens in firmware thej SFH 1.0 HID report descriptor is left untouched.

Co-authored-by: antheas <git@antheas.dev>
Signed-off-by: Jan200101 <sentrycraft123@gmail.com>