[redhat] New configs in drivers/scsi
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SCSI_LIB_KUNIT_TEST [=m]
 Type  : tristate
 Defined at drivers/scsi/Kconfig:70
   Prompt: KUnit tests for SCSI Mid Layer's scsi_lib
   Depends on: KUNIT [=m]
   Visible if: KUNIT [=m] && !KUNIT_ALL_TESTS [=m]
   Location:
     -> Device Drivers
       -> SCSI device support
         -> KUnit tests for SCSI Mid Layer's scsi_lib (SCSI_LIB_KUNIT_TEST [=m])

 Commit: 25a1f7a0a1fe (scsi: core: Add kunit tests for scsi_check_passthrough())

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>