[redhat] New configs in drivers/watchdog
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: LENOVO_SE10_WDT [=n]
 Type  : tristate
 Defined at drivers/watchdog/Kconfig:257
   Prompt: Lenovo SE10 Watchdog
   Depends on: WATCHDOG [=y] && (X86 [=y] && DMI [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> Watchdog Timer Support (WATCHDOG [=y])
         -> Lenovo SE10 Watchdog (LENOVO_SE10_WDT [=n])
 Selects: WATCHDOG_CORE [=y]

 Commit: 1f6602c8ed1e (watchdog: lenovo_se10_wdt: Watchdog driver for Lenovo SE10 platform)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>