[redhat] New configs in drivers/tty
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SERIAL_SC16IS7XX_CORE [=n]
 Type  : tristate
 Defined at drivers/tty/serial/Kconfig:1026
   Prompt: NXP SC16IS7xx UART support
   Depends on: TTY [=y] && HAS_IOMEM [=y]
   Location:
     -> Device Drivers
       -> Character devices
         -> Enable TTY (TTY [=y])
           -> Serial drivers
             -> NXP SC16IS7xx UART support (SERIAL_SC16IS7XX_CORE [=n])
 Selects: SERIAL_CORE [=y] && SERIAL_SC16IS7XX_SPI [=n] && SERIAL_SC16IS7XX_I2C [=n]

 Commit: 2c837a8a8f9f (sc16is7xx: spi interface is added)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>